refactor: organized folders to include decorator pattern
