--- src/appointments/domain/factories/appointment.abstract.ts
@@ -1,3 +1,0 @@
-export abstract class AppointmentAbstract {
-    public abstract calculateCost(): number;
-}
