+++ src/appointments/domain/factories/decorator/concrete/tyre-repair.ts
@@ -1,5 +1,5 @@
-import { AppointmentDecorator } from "./appointment-decorator";
-import { AppointmentAbstractFactory } from "./appointment-abstract.factory";
+import { AppointmentDecorator } from "../abstract/appointment-decorator";
+import { AppointmentAbstractFactory } from "../../appointment-abstract.factory";
 
 export class TyreRepair extends AppointmentDecorator {
     constructor(appointmentAbstractFactory: AppointmentAbstractFactory) {
